tedge mqtt pub - Fix a race when publishing QoS=0 messages (#323)
There was an issue when no connection could be established, in which
case "tedge mqtt pub" would return silently instead of printing an
error.

Errors are reported asynchronously. Publishing a message is async as
well, it just pushes the "publish request" to a channel. So there is no
way to notify the publisher immediatly that no connection exists. In
addition, we were not tracking whether the event loop actually did
process a Publish request of QoS=0 or not (we were only tracking
completed QoS=1 and QoS=2 messages). So after calling `publish` we were
immediatly terminating without any chance for the connection error to
propagate.

We now keep track about all pending messages, independent of the QoS
level, until they are known to be sent out by the event loop. We get
notified by the `Event::Outgoing(Outgoing::Publish(id))` whenever a
publish request was processed by the event loop.

In addition to that, we now keep separate counts for pending QoS=1 and
QoS=2 messages, where before we were just using one counter for both.
IMHO having separate counts is slightly more readable, but it can be
collapsed into one count again if there is desire.

We now expose a `MqttClient::has_pending() : bool` method, that we use
in the `tedge mqtt pub` command to wait until no further messages are
pending.

As calling `MqttClient::publish` will increase the
`pending_publish_count` also for QoS=0 messages, we will actually stay
in the loop that is checking for `has_pending` to drop to `false`. As
this is not happening in the case no connection exists (we do not
receive the `Event::Outgoing(Outgoing::Publish) event), the `errors`
channel has the chance to finally report the connection problem and
exit.

Co-authored-by: Didier Wenzek <didier.wenzek@acidalie.com>