[CIT-589] The start of batch is the event time of its first event
Previously, the start a new batch was computed after the processing time
(reduced by the delivery_jitter aka the maximum accepted delay),
while the event time was used to add an event in a batch.
* The first negative effect, was that the batch start date was not
  deterministic but dependent of the processing time.
* The main issue was that increasing the maximum accepted delay
  was reducing the span of events that could be added to a batch.
  Up to the point where each batch is formed only by its former point.

Now, the event time is used to define the span of a new batch and to
find a candidate batch for an event.
* The processing time is no more used to define a batch, but just to
  determine when there is no point to wait for more events.
* Increasing the maximum delay has no impact on grouping,
  but only on 1) the messages that are discarded because late
  and 2) an augmented delay to emit batches.

Futhermore, the