Implement Reactor main loop
This patch implements the running of all plugins in the reactor.

The plugins are run by moving the responsibility of communication of a
plugin with all others to a PluginTask, which implements the main loop
for handling messages from and to other plugins.

For each Plugin, one PluginTask is instantiated and run in parallel to
all other PluginTasks.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>