Route all messages over the CoreTask
This patch changes the layout so that messages from plugins to other
plugins are routed over the CoreTask.

This ensures that we do only have 3communication channels in each
PluginTask:

* One for CoreTask -> PluginTask
* One for PluginTask -> CoreTask
* One for Plugin -> PluginTask
(PluginTask -> Plugin is a function call: `Plugin::handle_message()` and
thus does not need a communication channel!)

and we thus do not run into the issue of having n^2 communication
channels for the communication between the PluginTask objects.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>