Impl Debug for Comms
This patch implements the `std::fmt::Debug` trait for the `Comms` type,
because when implementing plugins, we store a `Comms` object in a
`Plugin` type.
The `Plugin` type, though, needs to be `std::fmt::Debug` because we want
to `tracing::instrument` functions on it.

But because `Comms` is not `std::fmt::Debug`, we have to hand-implement
the `std::fmt::Debug` trait ob the respective `Plugin` type, which is
tedious and not a nice developer experience.

Thus, implement the trait on the `Comms` type, so the `Plugin` types
that store it have a higher chance of deriving the implementation of
the `std::fmt::Debug` trait.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>
Signed-off-by: Marcel Müller <m.mueller@ifm.com>
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>