Remove Clone bound on PluginDirectory trait
This patch removes the `Clone` trait bound from the `PluginDirectory`
trait, because we pass the `PluginDirectory` implementing type as a
reference to a `PluginBuilder` when it is used to instantiate a
`Plugin`. The `PluginBuilder` implementation can then pull the required
`Address`es from the instance of `PluginDirectory` and keep them, not
requiring the implementation to keep the whole `PluginDirectory`
object.

This makes the API simpler, and we can remove the `Clone` trait bound
because of this.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>
Signed-off-by: Marcel Müller <m.mueller@ifm.com>