Make Address/Reply{Receiver,Sender} contravariant
Previously the 'Address Types' were covariant with their generic type T. This
means that rustc understood `PhantomData<T>` to be 'equivalent' in
subtyping to `T`. This has implications for auto traits like `Send` and
`Sync`, because this means that the `Address Types` only implement those
if `T` does as well! But this is overly restrictive, the 'Address Types'
never contain a `T`! Nor do they ever touch it, hence we are using
`fn(T)` to tell rustc that it should treat it in a contravariant
fashion, meaning that in this case it is not dependent on `T` and always
implements `Send/Sync`.

Additionally, static tests were added to ensure this in the future.

This also removes some unsafe code, which makes this crate once again
free from it! 🎉

Signed-off-by: Marcel Müller <m.mueller@ifm.com>