Add missing tokio features
The tokio features "process" and "rt-multi-thread" were missing.
The issue did not come up when compiling the whole worktree, because
(I guess) other crates had these features enabled and thus "leaked" them
into the compile process for this one.
When building only this crate, it failed though because of the missing
features.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>