Add PluginDeclaration trait
This trait simplifies a lot of the message calls in the PluginBuilder.

Advantages: You only declare a list of handled messages _once_. This is
necessary since Rust does not allow 'queries' for types (aka "give me a
tuple of supported types" does not exist), and so the necessary crutch
is to declare them manually.

Disadvantages: You cannot pick at Runtime anymore what messages you
decide to support. This is considered to be a very niche use-case and
might actually be an anti-pattern.

Signed-off-by: Marcel Müller <m.mueller@ifm.com>