Introduce `ActorInstance` and `ActiveActor` handle types
Having these two types allows to create actors and to connect them
before running them. An actor instance wraps a state with a mailbox
but nothing is running yet. Using the maibox addresses of their peers,
actor instances can then be connected. When all the recipients of an
actor are connected, then the actor can be launched.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>