Add the Reactor trait
The motivation is to separate the two main behavior of an actor:
* Producing spontaneous messages
* Reacting to received messages by sending output messages

On start, an actor returns a (`Producer`, `Reactor`) pair.
These two objects might share some state, a network connection for
instance. In practice, some actors acts only as a source or as a
reactor. For that purpose, one can use `DevNull` that acts as an empty
source and silently consumes any received message.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>