Add the Task trait so an actor can spawn a task reacting to an input message
An example that motivates such trait objects is a "command" plugin
that receives requests to run commands in the background.
A long running should not block the execution of other commands.
Hence the necessity to spawn a Task as the result of an input message.

Another motivation is to simplify the Actor trait removing the
Producer associated type.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>