Use "systemd" crate instead of CLI of systemd
This patch changes the talking to the systemd process from using the CLI
interface of systemd to using the "systemd" crate.

The Error type had to be adapted to be able to bubble up errors from the
systemd crate interface.
If the talking to systemd failed (indicated by a `false` returned from
`systemd::daemon::pid_notify`, we return a WatchdogError::TalkingFailed.
There seems to be nothing more we can do.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>