Remove the `tedge` dependency on `UserManager`
The UserManager was used to run the `tedge` command as the `tedge` user
becomming `root` or `mosquitto` only when required.
In practice, this introduced more issues than benefits.
Notably, this forces to have a `tedge` user to run the command.
While switching to a different user is required only to create a
certificate under mosquitto.

The behavior is unchanged. Particularly, errors are ignored when giving
ownership of the certificate to mosquitto
(for instance, if the `tedge cert create` command is not run with `sudo`).

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>