Remove unused test
The test `create_certificate_as_root_should_switch_to_mosquitto` were
behind features and never triggered. As a consequence this test has not
been maintained and is no more aligned with the code base (the tedge config is
no more user dependent, there is --config-dir option, the tedge command
is no more running as tegde). Futhermore this test was using `sudo`.

Better to simply remove it.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>