Revert "Add cargo-outdated check with posting to PR"
So there were two issues with the patch.

First, jobs using `GITHUB_TOKEN` cannot run on PRs from forks in github.
This is an unfortunate restriction that keeps us from implementing such
a thing here.

Second reason (and maybe even more important) is the fact that having
outdated dependencies in the repository would result in comments on each
PR (if github would allow that), even though that PR did not even touch
the dependencies. A check for only newly added dependencies may not be
trivial (or even possible).

Of course the actual issue is that we have outdated dependencies, and
that should be resolved! But the idea of having a bot post to every PR
that there are outdated dependencies, even though the PRs do not even
touch dependencies, might confuse contributors and maintainers alike.

So the idea was to remove the cargo-outdated job. Rationale being that
we get PRs from dependabot for each outdated dependency anyways! So
there will be one PR for each outdated dependency as soon as a
dependency gets outdated! If a PR author introduces new dependencies, it
should be job of the author and the reviewer to make sure that the newly
introduced dependencies are up to date (or reasonably up to date). If
something slips by, dependabot will file a PR within one day to update
the dependency anyways!

This commit reverts 7abdd1e5537ffe53d1c15017fa5ea873b35c1b60 ("Add cargo-outdated check with posting to PR")

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>