Add support for request
A request group a request payload with a response sender.
So an actor handling `Request<Req,Res>` can directly send a response of
type `Res` to a requester that sent a request of type `Req`.

I'm not fully happy with proposal. On the positive side, this opens a way
to implement a request / response scheme. However, the responses are no
more observed as regular messages among the stream of output messages.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>