Make systemd watchdog backend only build for linux
This patch ensures that the systemd watchdog backend is only used when
building for linux.

Right now, there is no other watchdog implementation. Thus, there is a
"dummy_watchdog" module introduced that does nothing. This might be
removed and replaced by something specific, if we ever build
thin-edge.io for other targets, such as Apple, FreeBSD or others.

Suggested-by: Albin Suresh <albin.suresh@softwareag.com>
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>