Do not depend on fmt/clippy in long-running jobs
The build and test jobs take a really long time (approx 17 minutes for
the arm32 test job). We cannot afford to wait on cargo-fmt and
cargo-clippy with those, so remove the dependencies here to get a faster
overall time.

Instead, only depend on the "changes" job, so we only run if there are
rust/workflow changes.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>