Remove global mqtt dummy server
Before this patch there was one in-process mqtt server used for running
all tests. Tests had to be run in sequence because of this, adding up to
a huge test time.

This patch makes each test have its own broker, on an individual port,
to be able to run all tests in parallel.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>