Add what thin-edge.io provides for Service Providers
Reasoning:

A customer or partner has an existing IIoT project or an existing IIoT
platform/ environment. The project is mostly driven by the IT. The
project needs expertise and solutions to get access to data and
information form the field of industrial sensors and automation.

State that thin-edge.io solves that.

Reviewed-by: Matthias Betz <matthias.betz@ifm.com>
Signed-off-by: Matthias Betz <matthias.betz@ifm.com>
Signed-off-by: Marcel Müller <m.mueller@ifm.com>
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>