Add that the step from POC to production is easy
A customer has a focused problem and objective and needs a digital IIoT
solution. They want to avoid complex and time-consuming projects and
argues with fast return of investment. The solution has to scale, that
is the step from PoC to Rollout has to be easy and with minimal
friction.

Reviewed-by: Matthias Betz <matthias.betz@ifm.com>
Signed-off-by: Matthias Betz <matthias.betz@ifm.com>
Signed-off-by: Marcel Müller <m.mueller@ifm.com>
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>