Rewrite section on merging
In de3bcc7e16b69d753b132ebbbb4da8073cf18b74 ("Move "Repository maintenance" section to CONTRIBUTING.md")
we moved the "Repository maintenance" section to the CONTRIBUTING.md
file.
This patch moves the sections on merging and the evergreen-master
strategy.
The "Reviewing, addressing feedback, and merging" section was broken
apart for that and the "and merging" part was moved to the new section.

Fixes: de3bcc7e16b69d753b132ebbbb4da8073cf18b74 ("Move "Repository maintenance" section to CONTRIBUTING.md")
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>