Update debian maintainer scripts of tedge_agent
To enable self-update, when systemd is available on the device, on
upgrading, maintainer script doesn't stop tedge_agent service.
To complete upgrading, user need to restart a device.

There is a workaround regarding #DEPHELPER#
- postinst and postrm should not use #DEPHELPER#, however, cargo-deb
returns an error if #DEPHELPER# doesn't exist. To prevent this, stop
adding automatically the #DEPHELPER# code, and add them to maintainer
script directly.

Signed-off-by: Rina Fujino <18257209+rina23q@users.noreply.github.com>