Fix fragments tests to work with `tedge` installed
This test was passing because the pipeline did not have tedge installed.
But when tedge is installed, locally, the version field defaults to the
version of tedge on your system. This tests fixes this issue to use
tedge from cargo_bin. Additionally env! is used to get the current
version of the binaries, which can be used to compare to tedge
--version.

Signed-off-by: alexandru solomes <solo@softwareag.com>