Remove flockfile for tedge_agent before running --init in postint script
tedge_agent is supposed to run by the 'tedge' user, therefore, the
flockfile is owned by the 'tedge' user. However, due to the directory
creation of '/var/tedge', 'tedge_agent --init' must be run as root in
the postinst script. If the flockfile owned by 'tedge' user exists, the
'tedge_agent --init' will fail.

To fix this issue, the postinst script should remove the flock file if
it exists before the initialization.

Signed-off-by: Rina Fujino <18257209+rina23q@users.noreply.github.com>