Change the restart temporary file location from /run/tedge_agent to /tmp
tedge_agent creates `tedge_agent_restart` file in order to keep the
timestamp of restart operation executed. However, the file was supposed
to be created /run/tedge_agent directory, which is created by systemd.

In addition, to create a file directly under /run requires root.

To remove the implicit dependency on systemd, change the directory to
create the file from /run/tedge_agent to /tmp.

Signed-off-by: Rina Fujino <18257209+rina23q@users.noreply.github.com>