Fix the c8y custom operation status handling
A custom operation managed in /etc/tedge/operations/c8y could be marked
successful unconditionally and the status had been sent out before the
operation has completed.

To fix, now the Cumulocity converter holds a mqtt publisher and sends
executing and successful/failed message from the converter. Either
successful or failed is determined by the exit code (0 or not).

Also, stdout is reported when successful, and stderr for a failed case.

Signed-off-by: Rina Fujino <18257209+rina23q@users.noreply.github.com>