Declare child devices again if deleted in cloud
When deleting a child device from the cloud representation the supported
operations logic would re-create this child device but fail to notice it
is not available on the cloud any longer. Thus resulting in a child
device name prefixed by "MQTT Device".

This fix, changes the logic of initialising supported operations on
startup, to first requesting the cloud representation of child devices,
computing a diff between cloud and local, initialising as new child
devices anything that is available locally but is not seen on the cloud
and then proceeds to add supported operations for the child devices.

Signed-off-by: alexandru solomes <solo@softwareag.com>