Add output to `tedge cert` commands
Added output to `tedge cert create` and `tedge cert remove` commands:

- `tedge cert create`
  - `Certificate was successfully created.` if certificate was
    successfully created
- `tedge cert remove`
  - `Certificate was successfully removed.` if certificate was removed
  - `There is no certificate to remove.` if there is no certificate

The documentation was also updated to include the added output.

The change was made because I think it would be good for all the CLI
commands that could possibly be invoked by the human, to print
human-readable output describing what happened, even if just one
sentence, it's better than nothing.

E.g. `tedge cert remove` returned `Ok` for both `cert removed` and `no
cert` cases, but it had no output, so the user wouldn't know what
actually happened. Alternatively `no cert` case could return an error,
but I'd argue it makes more sense as a success case because the end
result is the same as with `cert removed` case: there is no cert and a
new one can be created.

As for the implementation, i separated the logic and printing between 2
impls, but e.g. `cli/certificate/show.rs` just does the printing in the
2nd impl, and if it's fine, then I could also move the print there and
avoid some boilerplate.

Lastly, sorry if I'm bikeshedding too hard 😅

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>