Fix a method to split SmartREST messages and improve failure reason
When receiving multiline SmartREST messages, splitting simply by `\n`
is not a right approach because a SmartREST message can also contain
`\n`, e.g. an argument of c8y_Command.

Also, in case of an operation failure, giving only the last line of
stderr is not easy enough for users to understand why the operation
failed. It should give also the whole context as much as possible.

Resolves #1687

Signed-off-by: Rina Fujino <18257209+rina23q@users.noreply.github.com>