Fix issues with `flock` preventing to launch a daemon even if not concurrently
Due to a combination of issues it was often required to delete the lock files
used by the mappers and the agent, especially when running these daemons
alternatively as tedge and root. This happens mostly by user mistakes,
but leads to a bad experience and has infected installation scripts.

The root cause is that these lock files (which are not deleted when the daemon
are killed) were created with no enough permissions to be used by
another user.

- The lock files are now created as writable by any users.
- To acquire the lock, the file is open without the `O_CREAT` option.
  This is to avoid issues in case fs.protected_regular=1
- On error, the message is now more explicit on the source of the error:
  a true concurrent access, or an issue with missing authorization to open the file.
  In the former case, this is a mistake to delete the lock file!

See: https://github.com/thin-edge/thin-edge.io/issues/1726

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>