Deprecate MessageBox::channel function
This function was too specialized and only usefull for testing.
Now tests will have to build message boxes exactly the same way as done
using actors. This adds a bit of burden on testing but reduce the number
of concepts as well as the gap between test and production code.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>