Add MQTT server authentication
This commit adds to our MQTT clients (tedge mqtt and mqtt_channel) the
capability to connect to secure MQTT brokers via TLS. For the broker
certificate to be successfully verified, this certificate has to be
signed by a CA certificate, and the CA certificate needs to be trusted
by the client.

To select a CA trusted certificate, new configuration settings,
`mqtt.client.ca_file` and `mqtt.client.ca_path` have been added.

Additionally the broker certificate needs to be X.509 v3 and have a
"X509v3 Subject Alternative Name" section containing its Common Name as
a DNS name. This behaviour happens upstream in rumqttc crate, and was
documented in [rumqtt#498].

[rumqtt#498]: https://github.com/bytebeamio/rumqtt/issues/498

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>