Simplify C8YHttpProxyActor after the simplified Actor trait
Previously, the actor trait has been implemented on the C8YHttpConfig
struct while the true actor was C8YHttpProxyActor. This trick has been
used to avoid passing the message box on all and every method call.
Now this trick is useless as the Actor trait let the message box
of an actor be owned by its state - aka self.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>