Add client authentication to `tedge mqtt` command
This commit adds 2 new config options: `mqtt.client.auth.certfile` and
`mqtt.client.auth.keyfile` to facilitate MQTT client authentication.

The client certificates need to be signed by a CA. To authorise these
clients, the certificate of the CA which signed client certificates
needs to be specified in mosquitto config with cafile or cadir options.
Description of these options can be found in [mosquitto.conf].

These new settings were added because we need different certificates to
identify a device on a remote broker, and to identify plugins or
extensions on a local broker.

Also rewrapped some doc comments to 80 columns in tedge_config_dto.rs

[mosquitto.conf]: https://mosquitto.org/man/mosquitto-conf-5.html

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>