c8y_api: Bumped csv to 1.2.1, fixed clippy lints
result-large-err lint was fixed by changing `SmartRestSerializerError`
to not contain `csv::IntoInnerError` which contains a writer and is
meant to offer a capability to recover from failure. But we simply just
returned this error to the caller, so the returned error was changed to
inner io::Error.

To be able to move the `io::Error` out of `csv::IntoInnerError`, csv
crate was upgraded to 1.2.1, which added `IntoInnerError::into_error`
function. This addition in patch version is not valid semver, so
dependency in `Cargo.toml` is specified as "1.2.1" (<= 1.2.1, < 2.0.0).

Additionally the error types of some functions changed. Functions which
returned `Result` but didn't actually return any errors were made
infallible and functions which returned a large error enums but could
actually only return a single variant which wrapped a smaller error, now
return this inner error.

Overall, these changes were made only to fix the lint, but error
handling could be simplified further, improving both readability and
performance, as described in #1767.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>