Add a Config type parameter to MessageSink
So a message sink can provides its config to connect the source.
Now a source and sink can be connected with `source.add_sink()`
without any knowledge of the config nor the internal sender used
by the sink (as it was the case with `sink.register_peer(config,&mut sender).

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>