Make components use MQTT authentication settings
This commit adds MQTT authentication support to the `mqtt_channel` crate
used by different plugins to access the local MQTT broker, as well as
changes the call sites in all the different plugins to start with MQTT
config partially filled by settings related to the current broker, e.g.
hostname, port, and authentication settings, which are the same for all
the plugins. After loading the pre-filled config, the client is then
free to provide settings for the client like subscriptions or last_will
messages, or even to override previously set broker settings.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>