Simplify MessageSink::add_mapped_input
The former implementation was taking a cast function transforming
DynSender, while the intent is to transform messages.
The new implementation takes a function that transforms an input
messages into a sequence of output messages. This transformation
is then used to connect two actors which messages are translated on the fly.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>