Refactor Robot tests to use MQTT authentication
To make sure that all the plugins can work properly with server and
client authentication enabled, the RobotFramework test suite setup was
edited to create a secure mosquitto listener on port 8883 with server
and client authentication enabled. This was achieved by adding
--secure/--no-secure options to bootstrap.sh script, with --secure being
the default.

The default port 1883 listener was not removed because it's created by
thin-edge with seemingly no easy way to disable editing mosquitto config
at runtime, thus it would require more source code changes which I've
decided are not worth it at the moment.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>