Add a connect_consumer method to the ServiceProvider trait
The motivation for this change is to get rid of the temporary use of
Option<DynSender<...>> in actor builders. Currently, these Option values are always
set to Some<sender>, because the previous version of the ServiceProvider trait
was forcing the caller to implement the ServiceConsumer trait
and indirectly to prematurely build the ServiceConsumer.
Now the constructor of an actor builder can directly connect a service provider.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>