tedge_config: bump toml to 0.7
There was a [bug] in version 0.5.11 of toml crate which caused
serialization to fail if a serialized struct had fields that serialized
to plain value keys after fields that serialized to tables (like vecs or
hashmaps).

In 0.7 this bug is solved and additionally `toml` crate is now a thin
wrapper over `toml-edit` crate, just providing serde compatibility.
`toml-edit` now is able to better edit toml without fully serializing
it, potentially helping with the partial serialization problem described
in this [comment].

[bug]: https://github.com/toml-rs/toml/issues/356
[comment]: https://github.com/thin-edge/thin-edge.io/issues/1812#issuecomment-1487349799

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>