Refactor TEdgeConfig to have a single source of truth
This builds on previous changes to TEdgeConfigDTO, adding auto generated documentation from
the doc comments, as well as a simpler means to access configuration via getters and update
values from both the type stored in the dto and a string representation.

This also migrates the tedge command to use the updated configuration, and reformats the
result of `tedge config list --doc` to be more colourful (and hopefully readable, too).

Signed-off-by: James Rhodes <jarhodes314@gmail.com>