tedge_config: add c8y.http, c8y.mqtt options
This commit adds `c8y.http` and `c8y.mqtt` config options intended to
override `c8y.url` option in HTTP and MQTT contexts respectively. This
is necessary for when HTTP and MQTT servers are present under different
domain names. These settings are optional and are derived from `c8y.url`
if not present.

As such, the use of `C8yUrlSetting` is discouraged, because it's not
clear whether it's used to connect to HTTP or MQTT. Because of current
limitations of `tedge_config` as well as timeliness considerations,
querying its value was not made deprecated, and will have to be done in
the future. The `c8y.url` can still be set, unset, or listed.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>