config: handle ports in c8y http/mqtt settings
Previously used for deserializing `ConnectUrl` type does not allow
specifying ports in host strings. To allow for deserializing with ports
optionally specified, as well as to know upfront what default port we
use in case it is not specified, and still handle possible parsing
errors, `HostPort` type was created, which uses a const type parameter
to specify which default port we expect.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>