Mark C8yUrlSetting as deprecated
Rust `#[deprecated]` attribute was used to mark `C8yUrlSetting`, with a
message to use `C8yHttpSetting` or `C8yMqttSetting` instead. This was
done because depending on the service we might want to use different
endpoints, e.g. when using custom domains and certificates which are at
the time of writing, supported by HTTP on Cumulocity, but not for MQTT.

Current necessary uses of the type, e.g. in `tedge config set` or `tedge
config list` commands, were marked with `#[allow(deprecated)]`
attribute, to not show up as warnings.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>