Fix error handling of RestartManagerActor
When an error occures, the RestartManagerActor manager must not give up
but return a proper error message the end user. The only exception to
that rule is when the actor fails to send messages to its peers. Then
logging the last error and giving up is the only way to go.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>