Remove NoParentDir variant from FileError
This variant was removed because it was very surface level and didn't
tell much about the actual source of the error.

Specifically, the error was used in `move_file` function, where it was
returned if we get `None` as the parent of the destination path. But,
`Path::parent()` only returns `None` if:

- path terminates in a root prefix ("/")
- is empty ("")

So, in both cases the issue is not actually that the path doesn't have a
parent, but the path is completely invalid as it would be invalid to
save a file under a name of either "/" or "".

Additionally, if we were to use a relative paths, setting the
destination to e.g. "destination.txt", its parent would become "", which
in turn would cause error when passed to `tokio::fs::create_dir_all`. If
forbidding relative paths was intended, a better name or a mention in
the comment would be sufficient.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>