Improve ConnectUrl parse error message
Previous error message was incorrect because there are other reasons for
parsing to fail than containing scheme or port, e.g. if the value is
empty, resulting in an message:

Provided URL: '' contains scheme or port
    Provided URL should contain only domain, eg: 'subdomain.cumulocity.com'.

The new message is now more generic, suggestion in the second line
should be sufficient to pinpoint the cause of the error by the user.

As a side note, we probably shouln't call that a URL if we only expect
some parts of the URL. Domain, address, or host would perhaps be better
names.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>