Fix bug in ConcurrentServerMessageBox
The bug in ConcurrentServerMessageBox::await_idle_processor caused
runtime shutdown requests to not have intended effect (closing the
actor) when the message box was processing in-progress requests.

This fix works for this specific case, but the message box can still
block in other places, e.g. when sending a response if the channel is
full. When this happens, if the receiver is not freed before the sender,
we block.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>