Allow log filtering with $RUST_LOG
`--debug` flag used to increase log verbosity is very noisy, as it
enables `trace` (lowest level) logs from all modules, as well as
dependencies.
On the contrary, `RUST_LOG` environment variable can be used to set
different logging levels on a per-module basis, allowing to e.g. set
`trace` for a module one is currently working on and `info` on the rest,
or limiting logging only to the current crate.

As such, `set_log_level` function was modified to not overide a
`$RUST_LOG` filter if it is present. If this variable is not present, it
sets log level globally as usual.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>