Sync file on atomically_write_file
The `atomically_write_file_async` and `atomically_write_file_sync`
functions correctly flush the temporary content before atomically moving it to its
destination. However, for the content to reach the disk,
these methods must also call `fsync` on the file and its directory.

This notably critical for the restart operation which use the file
system to resume the command when the system reboot.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>