Update mockito to 1.1
Previously used version of mockito (0.31) had an issue where a
destructor for objects created in spawned task would sometimes hang. It
caused the tests to hang as well because when returning `Result` from a
function (instead of `panic!`ing) the runtime would wait for all tasks
to finish.

This issue was fixed in 0.32.2, but the crate was bumped to the lastest
version as we were already using the latest version in another crate, so
this change also makes it so we only use 1 version of mockito.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>