Replace `?` in download tests with panics
`#[tokio::test]` reacts differently when a test returns a failure, and
differently when it panics: when returning error, it waits for other
tasks to finish, but when panicing, it immediately terminates the test,
which we want, because if any part of code panics, the test is failed
and the rest doesn't matter.

Moreover, in the unit tests instead of reporting errors at the end it's
better to panic because we can then see exactly where it was triggered
along with a message and a stacktrace.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>