Fix panic in tedge_config macro
`#[tedge_config(rename = "name-with-hyphens")]` panicked because of
hyphens in the rename field. This occured because a test that was being
generated to test that field did not remove hyphens from the identifier
name, e.g:

> error: proc macro panicked
> ...
>     = help: message:
`"example_value_can_be_deserialized_for_run_name-with-hyphens_example_0"`
is not a valid identifier

Was fixed by replacing hyphens with underscores in identifier name.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>