Don't log secrets when connecting to MQTT broker
PR #2233 introduced some `info!` logging statements to MQTT connection
code, so that the user could see what's happening.

However, this code printed out the contents of the AuthenticationConfig
struct which contained root certificate store and a client private key
if client authentication option was used.

The intent was to print out the certificate paths, but what was printed
was the contents.

Signed-off-by: root <marcel.guzik@inetum.com>