tedge-agent: move to new health topic
tedge-agent was updated to work with new health topics:

- it listens to health check command on a new topic:
  `te/[DEVICE_TOPIC_ID]/service/tedge-agent/cmd/health/check`
- it publishes its health status on a new topic:
  `te/[DEVICE_TOPIC_ID]/service/tedge-agent/status/health

*Converter*

However, to still be observable and work with other components and the
tests, the converter was modified to properly bridge new inputs and
output topics of tedge-agent:

- `tedge/health-check` and `tedge/health-check/SERVICE_NAME` are
  converted into `te/device/main///cmd/health/check` and
  `te/device/main/service/SERVICE_NAME/cmd/health/check` respectively,
  so that tedge-agent receives a health check command on a topic it
  expects
- `te/device/main/service/tedge-agent/status/health` and
  `te/device/CHILD_ID/service/tedge-agent/status/health` are converted
  into `tedge/health/tedge-agent` and
  `tedge/health/CHILD_ID/tedge-agent` respectively.
- in order not to cause a loop, a previous `tedge/health/SERVICE_NAME`
  -> `te/...` conversion was modified to not happen if SERVICE_NAME is
  "tedge-agent"

*tedge-api*

A new type, `ServiceHealthTopic` was introduced to denote health topics
of services. Currently it is a simple string wrapper, handling both new
and old health topics, but after all components are moved, it will
handle only new topics.

`ServiceTopicId` and `DeviceTopicId` were added as simple wrappers
of `EntityTopicId` that are confirmed to be either a service or a
device.

Because `te/device/main/service/tedge-agent/cmd/health/check` doesn't
have a command_id, an OperationType struct was modified to make
command_id optional.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>